fixed errors on several functions

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: jarregui <jarregui@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/25 16:20:20 by jarregui          #+#    #+#             */
-/*   Updated: 2022/11/02 19:05:50 by jarregui         ###   ########.fr       */
+/*   Updated: 2022/11/02 19:45:06 by jarregui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,24 +54,24 @@ int				ft_memcmp(const void *s1, const void *s2, unsigned int n); /* OK */
 void			*ft_memcpy(void *dst, const void *src, unsigned int n); /* OK */
 void			*ft_memmove(void *dst, const void *src, unsigned int len); /* OK */
 void			*ft_memset(void *b, int c, unsigned int len); /* OK */
-void			ft_putchar_fd(char c, int fd); /* MISSING --> P2*/
-void			ft_putendl_fd(char *s, int fd); /* MISSING --> P2*/
-void			ft_putnbr_fd(int n, int fd); /* MISSING --> P2*/
-void			ft_putstr_fd(char *s, int fd); /* MISSING --> P2*/
-char			**ft_split(char const *s, char c); /* MISSING --> P2*/
+void			ft_putchar_fd(char c, int fd); /* OK --> P2*/
+void			ft_putendl_fd(char *s, int fd); /* OK --> P2*/
+void			ft_putnbr_fd(int n, int fd); /* OK --> P2*/
+void			ft_putstr_fd(char *s, int fd); /* OK --> P2*/
+char			**ft_split(char const *s, char c); /* OK --> P2*/
 char			*ft_strchr(const char *s, int c); /* OK */
 char			*ft_strdup(const char *s1); /* OK */
-void			ft_striteri(char *s, void (*f)(unsigned int, char*)); /* MISSING --> P2*/
+void			ft_striteri(char *s, void (*f)(unsigned int, char*)); /* OK --> P2*/
 char			*ft_strjoin(char const *s1, char const *s2); /* OK --> P2*/
 unsigned int	ft_strlcat(char *dst, const char *src, unsigned int size); /* OK */
 unsigned int	ft_strlcpy(char *dst, const char *src, unsigned int size); /* OK */
 unsigned int	ft_strlen(const char *s); /* OK */
-char			*ft_strmapi(char const *s, char (*f)(unsigned int, char)); /* MISSING --> P2*/
+char			*ft_strmapi(char const *s, char (*f)(unsigned int, char)); /* OK --> P2*/
 int				ft_strncmp(const char *s1, const char *s2, unsigned int n); /* OK */
 char			*ft_strnstr(const char *haystack, const char *needle, unsigned int len); /* OK */
 char			*ft_strrchr(const char *s, int c); /* OK */
-char			*ft_strtrim(char const *s1, char const *set); /* MISSING --> P2*/
-char			*ft_substr(char const *s, unsigned int start, unsigned int  len); /* MISSING --> P2*/
+char			*ft_strtrim(char const *s1, char const *set); /* OK --> P2*/
+char			*ft_substr(char const *s, unsigned int start, size_t len); /* OK --> P2*/
 int				ft_tolower(int c); /* OK */
 int				ft_toupper(int c); /* OK */
